Genomic regions with the greatest evidence of transcription factor binding for the FOXA1 transcription factor (for demonstration purpose)

Genomic regions representing the greatest evidence of enrichment for the FOXA1 transcription factor (DCC accession: ENCFF000NAH) for regions chr1:249120200-249250621 and chr10:1-370100 from the Encyclopedia of DNA Elements (ENCODE) data (Dunham I et al. 2012).

data(A549_FOXA1_01_NarrowPeaks_partial)

Format

A GRanges containing one entry per genomic regions. Each row of GRanges has a name which represent the name of the experiment.

Source

The Encyclopedia of DNA Elements (ENCODE) (DCC accession: ENCFF000NAH)

References

Dunham I, Kundaje A, Aldred SF, et al. An integrated encyclopedia of DNA elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74.

See also

Examples

## Loading datasets
data(A549_FOSL2_01_NarrowPeaks_partial)
data(A549_FOSL2_01_Peaks_partial)
data(A549_FOXA1_01_NarrowPeaks_partial)
data(A549_FOXA1_01_Peaks_partial)

## Assigning experiment name to each row of the dataset.
## NarrowPeak and Peak datasets from the same experiment must
## have identical names.
names(A549_FOXA1_01_Peaks_partial) <- rep("FOXA1_01",
        length(A549_FOXA1_01_Peaks_partial))
names(A549_FOXA1_01_NarrowPeaks_partial) <- rep("FOXA1_01",
        length(A549_FOXA1_01_NarrowPeaks_partial))
names(A549_FOSL2_01_Peaks_partial) <-rep("FOSL2_01",
        length(A549_FOSL2_01_Peaks_partial))
names(A549_FOSL2_01_NarrowPeaks_partial) <- rep("FOSL2_01",
        length(A549_FOSL2_01_NarrowPeaks_partial))

## Calculating consensus regions for both chromosomes 1 and 10
## with a default region size of 300 bp (2 * extendingSize)
## which is not extended to include all genomic regions.
## A peak from both experiments must be present in a region to
## be retained as a consensus region.
chrList <- Seqinfo(c("chr1", "chr10"), c(249250621, 135534747), NA)
findConsensusPeakRegions(
    narrowPeaks = c(A549_FOXA1_01_NarrowPeaks_partial,
                            A549_FOSL2_01_NarrowPeaks_partial),
    peaks = c(A549_FOXA1_01_Peaks_partial,
                            A549_FOSL2_01_Peaks_partial),
    chrInfo = chrList,
    extendingSize = 150,
    expandToFitPeakRegion = FALSE,
    minNbrExp = 2,
    nbrThreads = 1)
#> $call
#> findConsensusPeakRegions(narrowPeaks = c(A549_FOXA1_01_NarrowPeaks_partial, 
#>     A549_FOSL2_01_NarrowPeaks_partial), peaks = c(A549_FOXA1_01_Peaks_partial, 
#>     A549_FOSL2_01_Peaks_partial), chrInfo = chrList, extendingSize = 150, 
#>     expandToFitPeakRegion = FALSE, minNbrExp = 2, nbrThreads = 1)
#> 
#> $consensusRanges
#> GRanges object with 11 ranges and 0 metadata columns:
#>        seqnames              ranges strand
#>           <Rle>           <IRanges>  <Rle>
#>    [1]     chr1 249120024-249120324      *
#>    [2]     chr1 249120434-249120734      *
#>    [3]     chr1 249123174-249123474      *
#>    [4]     chr1 249132140-249132440      *
#>    [5]     chr1 249153305-249153605      *
#>    [6]     chr1 249167314-249167614      *
#>    [7]     chr1 249167909-249168209      *
#>    [8]     chr1 249200068-249200368      *
#>    [9]    chr10       179474-179774      *
#>   [10]    chr10       313079-313379      *
#>   [11]    chr10       343155-343455      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> attr(,"class")
#> [1] "consensusRanges"