Sites with the greatest evidence of transcription factor binding for the NR3C1 transcription factor from ENCODE (DDC accession: ENCFF002CFR). For demonstration purpose.

Sites representing the greatest evidence of enrichment for the NR3C1 transcription factor (DCC accession: ENCFF002CFR) for regions chr2:40000000-50000000 and chr3:10000000-13000000 from the Encyclopedia of DNA Elements (ENCODE) data (Dunham I et al. 2012).

data(A549_NR3C1_CFR_Peaks_partial)

Format

A GRanges containing one entry per site. The peaks are surronded by ranges present in the dataset A549_NR3C1_CFR_NarrowPeaks_partial.

Source

The Encyclopedia of DNA Elements (ENCODE) (DCC accession: ENCFF002CFR)

Details

The peaks and ranges have been obtained using an optimal IDR analysis done on all replicates.

References

Dunham I, Kundaje A, Aldred SF, et al. An integrated encyclopedia of DNA elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74.

See also

Examples

## Loading datasets
data(A549_NR3C1_CFQ_NarrowPeaks_partial)
data(A549_NR3C1_CFQ_Peaks_partial)
data(A549_NR3C1_CFR_NarrowPeaks_partial)
data(A549_NR3C1_CFR_Peaks_partial)

## Assigning experiment name to each row of the dataset.
## NarrowPeak and Peak datasets from the same experiment must
## have identical names.
names(A549_NR3C1_CFQ_NarrowPeaks_partial) <- rep("NR3C1_CFQ",
                            length(A549_NR3C1_CFQ_NarrowPeaks_partial))
names(A549_NR3C1_CFQ_Peaks_partial) <- rep("NR3C1_CFQ",
                            length(A549_NR3C1_CFQ_Peaks_partial))
names(A549_NR3C1_CFR_NarrowPeaks_partial) <-rep("NR3C1_CFR",
                            length(A549_NR3C1_CFR_NarrowPeaks_partial))
names(A549_NR3C1_CFR_Peaks_partial) <- rep("NR3C1_CFR",
                            length(A549_NR3C1_CFR_Peaks_partial))

## Calculating consensus regions for chromosome 2
## with a default region size of 40 bp (2 * extendingSize)
## which is extended to include all genomic regions for the closest
## peak to the median position of all peaks included in the region (for
## each experiment).
## Peaks from both experiments must be present in a region to
## be retained as a consensus region.
chrList <- Seqinfo(c("chr2"), c(243199373), NA)
findConsensusPeakRegions(
    narrowPeaks = c(A549_NR3C1_CFQ_NarrowPeaks_partial,
                            A549_NR3C1_CFR_NarrowPeaks_partial),
    peaks = c(A549_NR3C1_CFQ_Peaks_partial,
                            A549_NR3C1_CFR_Peaks_partial),
    chrInfo = chrList,
    extendingSize = 20,
    expandToFitPeakRegion = TRUE,
    shrinkToFitPeakRegion = FALSE,
    minNbrExp = 2,
    nbrThreads = 1)
#> $call
#> findConsensusPeakRegions(narrowPeaks = c(A549_NR3C1_CFQ_NarrowPeaks_partial, 
#>     A549_NR3C1_CFR_NarrowPeaks_partial), peaks = c(A549_NR3C1_CFQ_Peaks_partial, 
#>     A549_NR3C1_CFR_Peaks_partial), chrInfo = chrList, extendingSize = 20, 
#>     expandToFitPeakRegion = TRUE, shrinkToFitPeakRegion = FALSE, 
#>     minNbrExp = 2, nbrThreads = 1)
#> 
#> $consensusRanges
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames            ranges strand
#>          <Rle>         <IRanges>  <Rle>
#>   [1]     chr2 40662093-49534300      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> attr(,"class")
#> [1] "consensusRanges"