Ranges associated to nucleosomes detected by the NOrMAL software using syntetic reads generated using a normal distribution. For demonstration purpose.

Ranges associated to nucleosomes detected by the NOrMAL software using syntetic reads generated using a normal distribution with a variance of 20 for regions chr1:10000-15000.

data(NOrMAL_nucleosome_ranges)

Format

A GRanges containing one entry per detected nucleosome. The ranges are surronding the nucleosomes present in the dataset NOrMAL_nucleosome_positions. The genomic ranges have been obtained by adding 73 bps on each side of the detected positions.

References

Polishko A, Ponts N, Le Roch KG and Lonardi S. 2012. NOrMAL: Accurate nucleosome positioning using a modified Gaussian mixture model. Bioinformatics 28 (12): 242-49.

See also

Examples

## Loading datasets
data(PING_nucleosome_positions)
data(PING_nucleosome_ranges)
data(NOrMAL_nucleosome_positions)
data(NOrMAL_nucleosome_ranges)
data(NucPosSimulator_nucleosome_positions)
data(NucPosSimulator_nucleosome_ranges)

## Assigning experiment name to each row of the dataset.
## Position and range datasets from the same sofware must
## have identical names.
names(PING_nucleosome_positions) <- rep("PING",
                            length(PING_nucleosome_positions))
names(PING_nucleosome_ranges) <- rep("PING",
                            length(PING_nucleosome_ranges))
names(NOrMAL_nucleosome_positions) <-rep("NOrMAL",
                            length(NOrMAL_nucleosome_positions))
names(NOrMAL_nucleosome_ranges) <- rep("NOrMAL",
                            length(NOrMAL_nucleosome_ranges))
names(NucPosSimulator_nucleosome_positions) <-rep("NucPosSimulator",
                            length(NucPosSimulator_nucleosome_positions))
names(NucPosSimulator_nucleosome_ranges) <- rep("NucPosSimulator",
                            length(NucPosSimulator_nucleosome_ranges))

## Calculating consensus regions for chromosome 1
## with a default region size of 30 bp (2 * extendingSize).
## Consensus regions are resized to include all genomic regions of
## included nucleosomes.
## Nucleosomes from at least 2 software must be present
## in a region to be retained as a consensus region.
chrList <- Seqinfo(c("chr1"), c(249250621), NA)
findConsensusPeakRegions(
    narrowPeaks = c(PING_nucleosome_ranges,
                            NOrMAL_nucleosome_ranges,
                            NucPosSimulator_nucleosome_ranges),
    peaks = c(PING_nucleosome_positions,
                            NOrMAL_nucleosome_positions,
                            NucPosSimulator_nucleosome_positions),
    chrInfo = chrList,
    extendingSize = 15,
    expandToFitPeakRegion = TRUE,
    shrinkToFitPeakRegion = TRUE,
    minNbrExp = 2,
    nbrThreads = 1)
#> $call
#> findConsensusPeakRegions(narrowPeaks = c(PING_nucleosome_ranges, 
#>     NOrMAL_nucleosome_ranges, NucPosSimulator_nucleosome_ranges), 
#>     peaks = c(PING_nucleosome_positions, NOrMAL_nucleosome_positions, 
#>         NucPosSimulator_nucleosome_positions), chrInfo = chrList, 
#>     extendingSize = 15, expandToFitPeakRegion = TRUE, shrinkToFitPeakRegion = TRUE, 
#>     minNbrExp = 2, nbrThreads = 1)
#> 
#> $consensusRanges
#> GRanges object with 27 ranges and 0 metadata columns:
#>        seqnames      ranges strand
#>           <Rle>   <IRanges>  <Rle>
#>    [1]     chr1 10002-10148      *
#>    [2]     chr1 10167-10314      *
#>    [3]     chr1 10334-10482      *
#>    [4]     chr1 10503-10649      *
#>    [5]     chr1 10670-10816      *
#>    ...      ...         ...    ...
#>   [23]     chr1 14177-14325      *
#>   [24]     chr1 14340-14488      *
#>   [25]     chr1 14506-14657      *
#>   [26]     chr1 14677-14824      *
#>   [27]     chr1 14842-14992      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> attr(,"class")
#> [1] "consensusRanges"