Nucleosome positions detected by the NucPosSimulator software using syntetic reads generated using a normal distribution. For demonstration purpose.

Nucleosome positions detected by the NucPosSimulator software using syntetic reads generated using a normal distribution with a variance of 20 for regions chr1:10000-15000.

data(NucPosSimulator_nucleosome_positions)

Format

A GRanges containing one entry per detected nucleosome. The surronding ranges associated to those nucleosomes are in the dataset NucPosSimulator_nucleosome_ranges.

References

Sch&ouml;pflin R, Teif VB, M&uuml;ller O, Weinberg C, Rippe K, and Wedemann G. 2013. Modeling nucleosome position distributions from experimental nucleosome positioning maps. Bioinformatics 29 (19): 2380-86.

See also

Examples

## Loading datasets
data(PING_nucleosome_positions)
data(PING_nucleosome_ranges)
data(NOrMAL_nucleosome_positions)
data(NOrMAL_nucleosome_ranges)
data(NucPosSimulator_nucleosome_positions)
data(NucPosSimulator_nucleosome_ranges)

## Assigning experiment name to each row of the dataset.
## Position and range datasets from the same sofware must
## have identical names.
names(PING_nucleosome_positions) <- rep("PING",
                            length(PING_nucleosome_positions))
names(PING_nucleosome_ranges) <- rep("PING",
                            length(PING_nucleosome_ranges))
names(NOrMAL_nucleosome_positions) <-rep("NOrMAL",
                            length(NOrMAL_nucleosome_positions))
names(NOrMAL_nucleosome_ranges) <- rep("NOrMAL",
                            length(NOrMAL_nucleosome_ranges))
names(NucPosSimulator_nucleosome_positions) <-rep("NucPosSimulator",
                            length(NucPosSimulator_nucleosome_positions))
names(NucPosSimulator_nucleosome_ranges) <- rep("NucPosSimulator",
                            length(NucPosSimulator_nucleosome_ranges))

## Calculating consensus regions for chromosome 1
## with a default region size of 50 bp (2 * extendingSize).
## The consensus regions are extended to include all genomic regions for
## all nucleosomes. However, if the consensus regions are larger than the
## genomic regions of the nucleosomes, the consensus regions are not
## shrinked.
## Nucleosomes from all software must be present in a region to
## be retained as a consensus region.
chrList <- Seqinfo(c("chr1"), c(249250621), NA)
findConsensusPeakRegions(
    narrowPeaks = c(PING_nucleosome_ranges,
                        NOrMAL_nucleosome_ranges,
                        NucPosSimulator_nucleosome_ranges),
    peaks = c(PING_nucleosome_positions,
                        NOrMAL_nucleosome_positions,
                        NucPosSimulator_nucleosome_positions),
    chrInfo = chrList,
    extendingSize = 25,
    expandToFitPeakRegion = TRUE,
    shrinkToFitPeakRegion = FALSE,
    minNbrExp = 3,
    nbrThreads = 1)
#> $call
#> findConsensusPeakRegions(narrowPeaks = c(PING_nucleosome_ranges, 
#>     NOrMAL_nucleosome_ranges, NucPosSimulator_nucleosome_ranges), 
#>     peaks = c(PING_nucleosome_positions, NOrMAL_nucleosome_positions, 
#>         NucPosSimulator_nucleosome_positions), chrInfo = chrList, 
#>     extendingSize = 25, expandToFitPeakRegion = TRUE, shrinkToFitPeakRegion = FALSE, 
#>     minNbrExp = 3, nbrThreads = 1)
#> 
#> $consensusRanges
#> GRanges object with 21 ranges and 0 metadata columns:
#>        seqnames      ranges strand
#>           <Rle>   <IRanges>  <Rle>
#>    [1]     chr1 10167-10314      *
#>    [2]     chr1 10334-10482      *
#>    [3]     chr1 10670-10816      *
#>    [4]     chr1 10836-10982      *
#>    [5]     chr1 11001-11148      *
#>    ...      ...         ...    ...
#>   [17]     chr1 13675-13821      *
#>   [18]     chr1 13842-13988      *
#>   [19]     chr1 14009-14156      *
#>   [20]     chr1 14677-14824      *
#>   [21]     chr1 14842-14992      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> attr(,"class")
#> [1] "consensusRanges"